+++ src/main/java/com/caoccao/javet/interop/engine/IJavetEngine.java
@@ -23,8 +23,6 @@ import com.caoccao.javet.interop.V8Runtime;
 
 @SuppressWarnings("unchecked")
 public interface IJavetEngine<R extends V8Runtime> extends IJavetClosable {
-    void gc();
-
     JavetEngineConfig getConfig();
 
     R getV8Runtime() throws JavetException;
@@ -38,4 +36,6 @@ public interface IJavetEngine<R extends V8Runtime> extends IJavetClosable {
     void resetContext() throws JavetException;
 
     void resetIsolate() throws JavetException;
+
+    void sendGCNotification();
 }